fix(statistics): adjust incorrect derivative tests to pass
Previously, the last two points in the second derivative test would
break the acceptable error. I concluded this was an implementation
error, but on further investigation, I think it's just an inherent flaw
in the algorithm. Please raise an issue if this is incorrect, but I
think it's just that the backwards difference algorithm is that much
less accurate than the central difference algorithm (10x, to be
precise). Here's a reimplementation in Desmos to show the error is not
specific to sciutil: <https://www.desmos.com/calculator/cfsxlfctgi>.
As such, I've adjusted the test to give much greater error margin for
the last two points.