docs(statistics): added remaining numeric derivatives to the Typst document
Added documentation for the traditional numeric derivatives to the Typst
document. This isn't done yet, I still need to write out the Taylor
Series expansions to prove the statements about error, but it's a start.
It also included the renaming of a few variables in the implementations
to align with the naming scheme of the documentation.