ci(general): fix variable misnaming causing cache misses
Caches were being generated by `cargo-deps` with hashes appended, but
other jobs were getting cache misses because they were trying to fetch
caches with empty hashes. This is because `cargo-deps`'s cache action
was using the correct variable name in its key, but `cargo-deps`'s
outputs (which other jobs relied on for the hash) weren't using the
correct variable name, so it passed an empty string.