fix(rounding)!: `Sign::Positive` now displays as `"+"`
Previously, `sciutil::rounding::digits::Sign` would display
`Sign::Positive` as an empty string and `Sign::Negative` as a string
with a hyphen (notably, this is not the Unicode character for a minus).
Now `Sign::Positive` will display as a string with a plus. This also
comes along with a small performance improvement for `Digits`'s
`Display` implementation, which now allocates the string with more
capacity.