feat(units): partial draft for composite physical units
This is a partial draft of my implementation for composite units. This
is an expansion of the unit system I initially made with
`sciutil::units::Float` to allow for more complex units to be expressed
--- the requirement that a unit's name and symbol being entirely known
at compile time made it impossible to composite units together, like in
representing "meters per second". This comes with a new trait,
`sciutil::units::FloatDisplay` and a few implementations. This commit
also comes with `sciutil::rounding::digits::UncertainDigits`, which
provides absolute uncertainty for values expressed with
`sciutil::rounding::digits::Digits`. This commit serves as more a
checkpoint than a feature addition, these new APIs are severely
undercooked. At the moment, I'm reading the
[SI Brochure](https://www.bipm.org/en/publications/si-brochure/) for
inspiration on more formal means of expressing units, because my
home-cooked method has several serious flaws (e.g., is the unit named
"kilogram-meters squared" referring to `(kg^2)(m^2)` or `(kg)(m^2)`?).