chore(units): partial draft for new unit types
This is a checkpoint commit for the new types used for representing
units. Assuming no unexpected barriers come up, this will replace the
still-experimental `sciutil::units::FloatDisplay`, and will likely come
with API changes for at least `sciutil::units::Float`. This commit is
marked as a chore because no changes have been made to the API or
implementation of any public code, this is strictly to get the new
source code into VCS.