chore(units): add arbitrary powers for units and symbols for unit chains to partial draft
Added `Power` to represent units raised to an `i32` power. Added
`Multiplied::flatten_symbols`, which (by default) uses
`Multiplied::flatten_units` to construct the symbol for the unit chain,
akin to `Unit::symbol`. Commit is marked as a chore because it is a
checkpoint commit for the ongoing unit overhaul, and does not in any way
impact a public API.