chore(units): rename draft unit handling module
Renamed the module for the ongoing unit handling overhaul from
`symbol_tree` to `composition`. I'm still not certain on this name,
though. This commit is marked as a chore because it is a checkpoint
commit for the ongoing unit overhaul, and does not in any way impact a
public API.