feat!: began integration of new unit system
New unit system has been integrated to the extent that all tests pass.
There's still lots of work to actually put it into use and to fix
outdated documentation, though. This commit (obviously) comes with many
extreme changes to or outright removal of many APIs. This includes for
the new unit system itself, e.g., `sciutil::composition::ValuedUnit` is
now called `Valued`, but there's also a new `sciutil::ValuedUnit` trait
that defines part of its behavior. That being said, `ValuedUnit` is on
the chopping block, I don't see a clear reason why that needs to be a
trait.